+++ sdk/communication/azure-communication-chat/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release History
 ## 1.1.0-beta.1 (Unreleased)
 - Added support for metadata in messages.
+- Added options class `TypingNotificationOptions` for setting `SenderDisplayName` of the notification sender.
 
 ## 1.0.0 (2021-06-15)
 ### New Features